🐛✨ Guess channel name based on all main body links
This kept breaking because it depends on finding a specific link class name in the HTML. Frustratingly, Twitch does not put the channel name in video urls anymore (i.e. https://www.twitch.tv/videos/##########) which would make this much easier. The new approach uses a heuristic that looks for a depth-1 link that occurs twice in the main page content. That will hopefully be more resilient to changes on Twitch's end.